Start DB and nginx proxy during "beach restart"
Fixes #36
